+++ packages/liblearn/index.js
@@ -1,0 +1,5 @@
+// Re-export classes for external use
+export { ExperienceLearner } from "./learner.js";
+export { ExperienceInjector } from "./injector.js";
+export { ExperienceStore } from "./store/experience.js";
+export { FeedbackIndex } from "./index/feedback.js";